refactor(auth schemas): refactor auth schemas file name
