+++ src/pages/login.tsx
@@ -10,7 +10,7 @@ import { useAuth } from 'contexts/AuthContext'
 import { useTheme } from 'contexts/ThemeContext'
 import { getCookie } from 'cookies-next'
 import { getAuthUser } from 'libs/auth/api'
-import { LoginFormData, loginSchema } from 'libs/auth/schemas/loginSchema'
+import { LoginFormData, loginSchema } from 'libs/auth/schemas'
 import { GetServerSideProps } from 'next'
 import Head from 'next/head'
 import Image from 'next/image'
